+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - **Minor**: feature additions
 - **Patch**: bug fixes, backward compatible model and function changes, etc.
 
+# v2.0.4 - 2024-03-22
+#### Changed
+* Updated addSecret() to allow multiple arguments
+```go
+func AddSecret(secrets ...string)
+```
+
 # v2.0.3 - 2020-05-21
 #### Changed
 * removed Gopkg.toml